requirements.txt
fastapi[standard]
python-dotenv
pydantic
mysql-connector
